refactor: move catalog pipeline to SDK (Phase 3)
Dissolve catalog.ts into SDK catalog/pipeline/ modules.
Move 6 pipeline modules + registry crawling.
Dynamic import for skill-discovery dependency.
CLI shims for backward compat.
