+++ packages/sdk/package.json
@@ -26,6 +26,9 @@
     "./providers/clients/*": "./src/providers/clients/*.ts",
     "./catalog": "./src/catalog/index.ts",
     "./catalog/*": "./src/catalog/*.ts",
+    "./catalog/pipeline": "./src/catalog/pipeline/index.ts",
+    "./catalog/pipeline/*": "./src/catalog/pipeline/*.ts",
+    "./catalog/availability": "./src/catalog/availability.ts",
     "./ui": "./src/ui/index.ts",
     "./ui/*": "./src/ui/*.ts"
   },
