+++ packages/sdk/src/catalog/index.ts
@@ -1,8 +1,9 @@
 // Catalog surface — pluggable storage for component indexes
 
+// Pipeline modules
+export { checkAllRepos, checkRepoAvailability } from './availability'
 export type { CatalogReader, CatalogStore, CatalogWriter } from './interface'
 export { CatalogManager, type CatalogManagerOpts, createCatalog } from './manager'
-
 export { NdjsonStore, type NdjsonStoreOptions } from './ndjson/index'
 export {
   appendNdjsonFile,
@@ -11,7 +12,7 @@ export {
   serializeNdjson,
   writeNdjsonFile,
 } from './ndjson/io'
-
+export * as pipeline from './pipeline/index'
 export { SqliteStore } from './sqlite/index'
 export type {
   AnalysisFields,
