+++ packages/sdk/src/catalog/pipeline/index.ts
@@ -1,0 +1,11 @@
+// Catalog pipeline — discovery, download, reconciliation, staleness, crawling
+
+export * from './compute'
+export * from './crawl'
+export * from './discover'
+export * from './download'
+export * from './io'
+export * from './manifest'
+export * from './reconcile'
+export * from './stale'
+export * from './types'
